refactor: 사전 식물 정보 서버 컴포넌트로 변경
