+++ src/components/@common/Tag/Tag.style.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import styled from 'styled-components';
 import theme from 'style/theme.style';
 import { TagVariantType } from '.';
