+++ src/components/dictionaryPlant/DictionaryPlantContent/index.tsx
@@ -1,19 +1,13 @@
-'use client';
-
 import type { DictionaryPlantExtendCycles } from 'types/dictionaryPlant';
-import { useRouter } from 'next/navigation';
-import { useSetRecoilState } from 'recoil';
 import SeeMoreContentBox from 'components/@common/SeeMoreContentBox';
 import SvgIcons from 'components/@common/SvgIcons/SvgFill';
-import SvgFill from 'components/@common/SvgIcons/SvgFill';
-import SvgStroke from 'components/@common/SvgIcons/SvgStroke';
+import GardenButton from 'components/dictionaryPlant/GardenButton';
 import TagBox from 'components/dictionaryPlant/TagBox';
 import TagSwitch from 'components/dictionaryPlant/TagSwitch';
 import {
   Accent,
   ContentBox,
   FamilyName,
-  GardenButton,
   HeaderBox,
   InformationTagBox,
   ManageInfoBox,
@@ -21,9 +15,8 @@ import {
   PropBox,
   PropsBox,
 } from './DictionaryPlantContent.style';
-import { selectedDictionaryPlantState } from 'store/atoms/garden';
 import parseTemperature from 'utils/parseTemperature';
-import { NO_INFORMATION, URL_PATH } from 'constants/index';
+import { NO_INFORMATION } from 'constants/index';
 import theme from 'style/theme.style';
 
 const DictionaryPlantContent = (props: DictionaryPlantExtendCycles) => {
@@ -44,18 +37,6 @@ const DictionaryPlantContent = (props: DictionaryPlantExtendCycles) => {
     waterOptions,
   } = props;
 
-  const setSelectedDictionaryPlant = useSetRecoilState(selectedDictionaryPlantState);
-  const router = useRouter();
-
-  const goFilteredGarden = () => {
-    setSelectedDictionaryPlant({
-      id,
-      name,
-      image,
-    });
-    router.push(URL_PATH.garden);
-  };
-
   const place = postingPlace.map((position, idx) => {
     const text = position === NO_INFORMATION ? '제공된 정보가 없어요😢' : position;
     return <TagBox.Content key={idx}>{text}</TagBox.Content>;
@@ -70,15 +51,7 @@ const DictionaryPlantContent = (props: DictionaryPlantExtendCycles) => {
         <div>
           <Name>{name}</Name>
           <FamilyName>{familyName}</FamilyName>
-          <GardenButton type="button" onClick={goFilteredGarden} aria-label="모두의 정원">
-            <SvgStroke
-              color={theme.color.fontPrimaryForBackground}
-              size={16}
-              icon="bulletin-board-line"
-            />
-            모두의 정원
-            <SvgFill color={theme.color.sub} size={12} icon="line-arrow-right" />
-          </GardenButton>
+          <GardenButton id={id} name={name} image={image} />
         </div>
       </HeaderBox>
       <ContentBox>
