+++ src/components/dictionaryPlant/TagBox/ingredients/Main.style.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { styled } from 'styled-components';
 
 interface FlexBoxProps {
