+++ src/components/dictionaryPlant/TagBox/ingredients/Title.style.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { styled } from 'styled-components';
 import theme from 'style/theme.style';
