+++ src/utils/getFilteredChildren.ts
@@ -12,6 +12,7 @@ const getFilteredChildren = (
   childrenProp: React.ReactNode,
   count?: number
 ) => {
+  console.log(childrenProp);
   const children = Children.toArray(childrenProp);
 
   return children
